Build liblan7430-config-lib.so as a versioned shared library
Add VERSION and SOVERSION to set_target_properties() in the library
CMakeLists.txt so that liblan7430-config-lib.so is a symbolic link to a
versioned .so library file.

This will allow multiple versions of the library to be installed on the
system to support API revisions. Also, fixes build errors with the Yocto
Project & bitbake which require versioned .so library files.

Was:

build/lib/
└── liblan7430-config-lib.so

Now:

build/lib/
├── liblan7430-config-lib.so -> liblan7430-config-lib.so.1
├── liblan7430-config-lib.so.1 -> liblan7430-config-lib.so.1.1.0
└── liblan7430-config-lib.so.1.1.0

Signed-off-by: Geoff Parker <geoffrey.parker@arthrex.com>